Fix RISC-V `sys_fstat`
