+++ kernel/src/arch/x86/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod cpu;
 pub mod signal;
+pub mod stat;
